+++ src/shared/services/alert/alert.service.ts
@@ -4,6 +4,7 @@ import { NavigationStart, Router } from '@angular/router';
 import 'rxjs/operator/publish';
 import { ConnectableObservable, Observable } from 'rxjs/Rx';
 import { Alert } from '../../domain/alert/alert.model';
+import { AlertLevel } from '../../domain';
 
 @Injectable()
 export class AlertService {
@@ -11,15 +12,24 @@ export class AlertService {
   private successObserver: Observer<Alert>;
   private errorObserver: Observer<Alert>;
   private warningObserver: Observer<Alert>;
+  // TODO: Remove
+  private alertObserver: Observer<Alert>;
 
   // the preserved messages
   preservedError = '';
   preservedWarning = '';
   preservedSuccess = '';
 
+  // TODO: Remove
+  message: string;
+  level: AlertLevel;
+
   successes: ConnectableObservable<Alert>;
   errors: ConnectableObservable<Alert>;
   warnings: ConnectableObservable<Alert>;
+  // TODO: Remove
+  alerts: ConnectableObservable<Alert>;
+
   private preserveAlerts = false;
 
   constructor(private router: Router) {
@@ -38,6 +48,12 @@ export class AlertService {
       .create(observer => this.warningObserver = observer)
       .publish();
     this.warnings.connect();
+
+    // TODO: Remove
+    this.alerts = Observable
+      .create(observer => this.alertObserver = observer)
+      .publish();
+    this.alerts.connect();
 
     this.router
       .events
@@ -119,5 +135,16 @@ export class AlertService {
     } else {
       return '';
     }
+  }
+
+  // TODO: Remove
+  push(msgObject) {
+    this.message = msgObject.message;
+    this.level = msgObject.level;
+
+    this.alertObserver.next({
+      level: this.level,
+      message: this.message
+    });
   }
 }