+++ RELEASE-NOTES.md
@@ -1,8 +1,11 @@
 ## RELEASE NOTES
 
-### Version 2.69.2-alert-service-update
+### Version 2.70.1-alert-service-update
 **EUI-3229** Task event completion - part 2
 **EUI-3232** CCD UI toolkit Alert service message improvements
+
+### Version 2.70.0-fix-alert-service
+Fix breaking changes of Alert Service
 
 ### Version 2.69.1-fix-replace-document-issue
 **EUI-2243** Fix upload replacement document issue